+++ src/core/albums/dto/create-album.dto.ts
