+++ src/core/albums/exceptions/http-exceptions.ts
