+++ src/core/artists/artists.controller.ts
