+++ src/core/artists/dto/create-artist.dto.ts
