+++ src/core/artists/dto/get-artist.dto.ts
