+++ src/core/artists/dto/update-artist.dto.ts
