+++ src/core/artists/exceptions/http-exceptions.ts
