+++ src/core/favorites/dto/create-favorite.dto.ts
