+++ src/core/favorites/dto/update-favorite.dto.ts
