+++ src/core/favorites/entities/favorite.entity.ts
