+++ src/core/favorites/exceptions/http-exceptions.ts
