+++ src/core/favorites/favorites.controller.ts
